Updated lockfile message for "composer install" command
The behaviour of the lockfile was a little confusing to me.
I assumed if I added a new dependency I should run
'composer install' again. It was not immediately obvious
from the output of the install command what was going on
and I was confused why my new package had not been installed.

I understand that a better fix would be to compare
composer.json to the lockfile, but until that has been
implemented this will maybe help some newbies out.