Changed some variable names, continue 2, locked install changed
Changed some some instances of the $package variable name to be
able to be consistent in all branches of the "add to request"
section of code.

Updated an instance of "break" to "continue 2" as it was resulting
in the same package getting both update() and install() called
on request. Seems like this is probably not needed, should be
just one or the other? (could be way wrong on this)

Gave "locked" branch of the update/locked/install code some
logic to handle internally installed packages.