[version-invalidation] Allow invalidation bad versions through git-notes
Implements invalidation for git and github.

We currently have a problem with old versions that have faulty composer.json
which kill the solver. We need a way to mark those versions as bad, so they
get skipped by the repository.

This is only a problem for tags, as we cannot fix them. The way this is
implemented is using git-notes to attach a note to the tagged commit. If this
note contains a line with the string "invalid" (the note is namespaced to
"composer"), that tag is skipped.