Introduce "title" property to schema
Introduce a new property "title" which can contain a concise human-readable
title for the package. Backwards compatibility: If a "title" property exists,
it is used as a label in listings. If it's empty, the package name could be
used as a fallback. But in the end it's up to the application which
information is displayed.

Fixes: #1140