Play nicely with external software already using Composer autoloader
Check to see if ComposerAutoloaderInit already exists. If it does, get the
loader from there. Otherwise, require the autoload_real.php as before.

Primary use case is phpunit. If phpunit is installed globally via Composer
it is not able to be used to run on a project that also uses Composer.