Added option to only show available packages
Added the --available (-a) option to the show command to only list
the available packages, similar to the --installed and --platform
options.

Additionally changed the output formatting when limiting the
package result to remove the hierarchy when only one type is being
showed. This facilitates parsing of a list of packages (for example
for shell scripting and completion).