Clarify custom installer names/interfaces
"i.e." is an abbreviation for "id est", meaning "that is". The intention here is probably "e.g.", "exempli gratia", meaning "for example".

Also gave non-`Composer\` example path and added the name of the interface an installer should implement (it's mentioned later in the article, but it's useful to mention it early).