Allow autoloading empty paths
This issue manifested itself with PSR-0 autoload defined for a script
in the root package. That script could not be run because the file was
not PSR-0 autoloadable.

The issue is that the ClassLoader always does "$dir/$classPath", which
if $dir is empty will make an absolute path which obviously does not exist.

This change modifies the class loader to allow for empty paths.