Do not hardcode vendor dir exclusion on archive.
For one thing this wouldn't have worked for any custom installers anyway
which can write installed code to other places. This will now allow one
to use composer archive on a clean code checkout to build an archive as
we are used to. Or on one that had composer install run to build an
archive that can be used for deployment which includes the vendors.