Use default description and license from CLI args
When running `composer init` with `--description` and `--license` arguments
on the command line they are later suggested as defaults during the
interactive flow. However when you press Enter (to use the default suggesstion)
Composer does not use it, but instead skip them entirely from the `composer.json`
generation.

This change provides a default argument not only to `DialogHelper::getQuestion()`,
but also to `DialogHelper::ask()`.