Format the possible version formats as a table
Note I: This syntax is supported by MarkdownExtra, which both getcomposer
and github are using.

Note II: This patch currently breaks the PDF, because pandoc does not like
non-standard markdown tables. Ideas for fixing this appreciated.

Note III: The idea for this patch came up a few weeks ago on IRC. We
agreed back then that a table would be a good idea.

Note IIII: This patch creates a stability section which opens the door for
finally documenting how stability works in the composer docs.