PlatformRepository now takes an array of packages to override.
These overrides are on the form
array($packageName => $version)

These "virtual" packages are added instead of the real ones,
with the version specified. Note that it is a version and not
a constraint so you cannot pass array('ext-mongo' => '*').

Also there is currently no limit what packages can be overriden.