Added a check whether scope isolated function `includeFile` exists before declaring it
This simple check enables integration of several projects that use Composer as their autoload manager by checking whether a global method `includeFile` exists or not, thus avoiding potential PHP fatal error. If one includes files of a project that already used Composer to manage dependencies (or autoloading), a clash is inevitable when it comes to defining the `includeFile` method - thus raising a PHP fatal error. This patch checks whether the method is already defined.