Require only once `files` (defined in the schema).
Imagine the following scenario:

  * 2 projects: A and B.
  * Both have a `autoload.files` directive in there respective
    `composer.json` files: A.a and B.b.
  * A.a loads B.b.

Composer will load files in this order: A.a and B.b. But because A.a
already did the load of B.b, when Composer does that, there is an error
(“<entity> already exists” for instance).

While, in appareance, it seems a bug that A.a loads B.b, according to
the context, it makes sense. For instance, we install dependencies of a
project with Composer but we don't use its autoloader. In this case, we
need a specific autoloader, the one from A for instance.

From the point of view of A, it is difficult to avoid the load of B.b
because Composer did not load it yet.

So this patch does one simple think: It `require_once` the
`autoload.files` instead of `require`.