Try to symlinks on Windows systems, too
The symlink() function supports Windows since PHP 5.3. Most current Windows installations use NTFS, which supports symlinks.

The problem with these wrapper scripts is that they don't work as expected in some cases, for example if the bin file is a jar file.

In this case, on Windows, you will end up calling "java -jar vendor/bin/example.jar", with example.jar really being an sh script.

So we should try to create symlinks on Windows, too. In case it doesn't work, there is still a fallback.