Always optimize autoloader when classmap is authoritative
When the classmap is authoritative, a non-optimized classmap
will mean the autoloader does not load some of the files it is
supposed to. composer require/remove have no way of generating
optimized classmaps and thus would break an installation using
composer in authoritative mode. Thus, there is reason not to use
optimization when classmap-authoritative is enabled.

Fixes #4244