Allow optionally disabling the static autoloader
The static autoloader presents a problem for environments where pre-PHP
5.6 syntax compatibility is required. Many projects have automated CI
that runs a linter across all files in the repository to prevent
developers from accidentally introdcing incompatible syntax. And while
the static autoloader properly guards itself on lower PHP versions, it
will still trip up linters. Given the number of bugs filed about this,
it appears to be a common setup: #5255, #5316, #5324, and #5407.

This adds a --no-static option to the dump-autoload command, which will
disable the static autoloader by updating autoload_real.php to not look
for the file, and removes autoload_static.php if it exists.