Added reference to logics/extcheck
logics/extcheck leveraging composer's functionality to require PHP extensions. Unfortunately in the way it is currently stands most packages on Packagist fail to require extensions used. Because of this installation of such packages on minimal PHP installations succeeds but then fails at run time. I think it is reasonable to mention existence of such tool on Composer site.