Warn about possible retag of version
In the inadvisable event that a package has been retagged, it’s possible
to be in a state where the versions are the same, but the underlying ref
is different. We report that there has been a version variation, but the
output is not helpful (`From 1.1.3 to 1.1.3`, which just causes more
confusion).

Since this is an exceptional case, it makes sense to give some insight
into why things may appear strange.