mute first cd in win proxy stub for sh shell, fixes #6567
in the sh script, the first cd command did output the directory it cd'ed
into did output the path itself.

fix for me is to redirect it's standard output to /dev/null

using git bash on a windows 10 box.

w/o this fix, the workaround is to invoke the .bat file which is created
as well (but this is not so handy as a non windows user for me).