Take only displayed packages into account to determine column width
The ShowCommand was taking all packages into account when determining the
max lengths used to display the width of each column. This was causing
unnecessary hiding of columns in case of using a longer name or a longer
version in a different package.
This was especially visible when using the outdated command, as it applies
filtering by default.