Generate same package mutual exclusion ("spme") rules on-demand to reduce memory.
The vast majority of the total objects Composer previously generated express the
idea that package A 1.0 conflicts with package A 1.1, and 1.2, and ...
Since the number of these rules expands as n*n for the number of available
versions of each package, in many real-world evaluations these rule objects may
dominate Composer's total memory utilization.

Instead of doing that, this modification to the solver only instantiates those
rules as needed while propagating decisions. Intuitively, when we decide A 1.0,
all other packages providing A can be decided against. That's essentially what
this does by newing up conflict unit rules for the other providers of A into
RuleWatchGraph::propagateLiteral() through the RuleWatchChain iterator that
propagateLiteral() pumps from. The decision is made and the SPME rule has served
its purpose; its refcount drops to zero and the memory is reclaimed.

A side-effect of not generating the full compliment of rules upfront is that
watch chains that would only contain SPME rules are not added to the RuleWatchGraph
at all, since the watch graph is built based on the contents of the RuleSet.
This is compensated for by creating that too on demand in propagateLiteral() when
necessary.