Add a small delay to the CacheClear command
Yes, this is insane. But it might make sense. 

There is no real benefit of clearing the local cache. But people (including myself) like to do it because we run in to errors. 
The most common error (for me) is that I have to wait until pagagist systems are updated. So having this "forced wait" will actually "fix" my problem.