StreamContextFactory: set the proxy protocol always to tcp://
The traffic between the proxy and the client is in some sense
unencrypted. The client should send a HTTP CONNECT request over
plaintext, which would contain the hostname it wants to connect to. The
proxy server would grant the request, open a TCP connection to this
hostname and then tunnel all traffic between the client and the remote
host. This mechanism is very simple. The proxy does not do any
encryption processing and does not see the data (unless it is a MITM
proxy, which is another story).

This is the ordinary, traditional way HTTP proxies work when it comes to
HTTPS tunnelling. Therefore, every client has to be explicitly
configured to use this proxy so that it would send the CONNECT request.
Simple redirection (e.g. via iptables) does not work.

Fixes: #8255