$home -> $COMPOSER_HOME
the `$home` variable referenced in the conifguration docs for data- and
cache-dir is a reference to an internal PHP variable that contains the
composer home directory.

as on the documentation page it is not visible where that variable comes
from (the document uses only shell/environment variables otherwise) these
two places are hard to read/understand.

the exact meaning of `$home` (lowercase) is the composer-home (or just
home) global configuration setting (composer home is for all user-wide
composer invocations, to store data in the filesystem namespace of the
user executing composer).

that meaning is better documented with the `$COMPOSER_HOME` environment
variable (which when it does not exists the default value is well
documented in [03-cli.md#composer-home].

replace `$home` with `$COMPOSER_HOME`.

additionally small typo fix on "rollback" which is "roll back".

discussion: #10363
[03-cli.md#composer-home]: doc/03-cli.md#composer-home