Explicit the version where script-aliases was included
Since this was included in a minor version, it's nice to tell the user if this feature is available on the version they're running or if they need to upgrade. If it's not available, the only hint (besides the missing aliases on `composer list`) is the mild complaint at `composer validate`.