Update reactphp/promise for PHP 8.5 compatibility
At this moment, nearly every single `composer install` on PHP 8.5 is failing on a deprecation notice coming from the `react/promise` package.

To mitigate this, the ReactPHP team has just released a new version and I'd like to suggest for Composer to update to that version and release a new version of Composer itself to unblock early adopters from testing with PHP 8.5.

Note: other than fixing the deprecation notice, the `3.3.0` release does not contain any user-facing changes.

Ref:
* https://github.com/reactphp/promise/pull/264
* https://github.com/reactphp/promise/releases/tag/v3.3.0
* https://github.com/reactphp/promise/compare/v3.2.0...v3.3.0